// src/app/modules/configuration/finance/models/fee-terms-data-view-model.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * FeeTermsViewModel
 */
export interface FeeTermsDataViewModel {
    /**
     * Gets or Sets ShortName
     */
    shortName?: string;
    /**
     * Gets or Sets DisplayName
     */
    displayName?: string;
    /**
     * Gets or Sets Code
     */
    code?: string;
    /**
     * Gets or Sets Description
     */
    description?: string;
    /**
     * Gets or Sets StartDate
     */
    startDate?: Date | string;
    /**
     * Gets or Sets EndDate
     */
    endDate?: Date | string;
    /**
     * Gets or Sets InvoiceGenerationDate
     */
    invoiceGenerationDate?: Date | string;
    /**
     * Gets or Sets InvoiceIssueDate
     */
    invoiceIssueDate?: Date | string;
    /**
     * Gets or Sets InvoiceDueDate
     */
    invoiceDueDate?: Date | string;
    /**
     * Gets or Sets InvoiceGenerationOptionTypeId
     */
    invoiceGenerationOptionTypeId?: number;
}