// src/app/modules/configuration/finance/models/fee-terms-view-model.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * FeeTermsViewModel
 */
export interface FeeTermsViewModel {    
    /**
     * Gets or Sets ShortName
     */
    shortName?: string;
    /**
     * Gets or Sets DisplayName
     */
    displayName?: string;
    /**
     * Gets or Sets Code
     */
    code?: string;
    /**
     * Gets or Sets Description
     */
    description?: string;
    /**
     * Gets or Sets StartDate
     */
    startDate?: Date;
    /**
     * Gets or Sets EndDate
     */
    endDate?: Date;
    /**
     * Gets or Sets InvoiceGenerationDate
     */
    invoiceGenerationDate?: Date;
    /**
     * Gets or Sets InvoiceIssueDate
     */
    invoiceIssueDate?: Date;
    /**
     * Gets or Sets InvoiceDueDate
     */
    invoiceDueDate?: Date;
    /**
     * Gets or Sets InvoiceGenerationOptionTypeId
     */
    invoiceGenerationOptionTypeId?: number;
    id?: string;
}